+++ Source/SwiftLintFramework/Extensions/SwiftSyntax/ModifierListSyntax+Extensions.swift
@@ -1,0 +1,8 @@
+import SwiftSyntax
+
+extension ModifierListSyntax {
+    /// Convenience variable to collect all modifiers from a declaration syntax node
+    var names: [String] {
+        map { $0.name.text.trimmed }
+    }
+}