+++ Source/SwiftLintFramework/Extensions/SyntaxVisitor+SwiftLint.swift
@@ -1,19 +1,19 @@
 import Foundation
 import SwiftSyntax
 
-// workaround for https://bugs.swift.org/browse/SR-10121 so we can use `Self` in a closure
-protocol SwiftLintSyntaxVisitor: SyntaxVisitor {}
+/// Workaround for https://bugs.swift.org/browse/SR-10121 so we can use `Self` in a closure
+public protocol SwiftLintSyntaxVisitor: SyntaxVisitor {}
 extension SyntaxVisitor: SwiftLintSyntaxVisitor {}
 
 extension SwiftLintSyntaxVisitor {
-    func walk<T>(tree: SourceFileSyntax, handler: (Self) -> T) -> T {
+    func walk<SyntaxType: SyntaxProtocol, T>(node: SyntaxType, handler: (Self) -> T) -> T {
         #if DEBUG
         // workaround for stack overflow when running in debug
         // https://bugs.swift.org/browse/SR-11170
         let lock = NSLock()
         let work = DispatchWorkItem {
             lock.lock()
-            self.walk(tree)
+            self.walk(node)
             lock.unlock()
         }
         let thread = Thread {
@@ -31,7 +31,7 @@ extension SwiftLintSyntaxVisitor {
 
         return handler(self)
         #else
-        walk(tree)
+        walk(node)
         return handler(self)
         #endif
     }
@@ -41,6 +41,6 @@ extension SwiftLintSyntaxVisitor {
             return []
         }
 
-        return walk(tree: syntaxTree, handler: handler)
+        return walk(node: syntaxTree, handler: handler)
     }
 }